+++ src/dune_engine/execution_parameters.mli
@@ -31,6 +31,8 @@ val set_dune_version : Dune_lang.Syntax.Version.t -> t -> t
 
 val set_swallow_stdout_on_success : bool -> t -> t
 
+val set_fail_on_non_empty_stderr : bool -> t -> t
+
 (** As configured by [init] *)
 val default : t Memo.Build.t
 
@@ -43,6 +45,8 @@ val should_remove_write_permissions_on_generated_files : t -> bool
 val should_expand_aliases_when_sandboxing : t -> bool
 
 val swallow_stdout_on_success : t -> bool
+
+val fail_on_non_empty_stderr : t -> bool
 
 (** {1 Initialisation} *)